==> multi-parameter-estimation/post-processing/1.1_sum_scale_and_chunk.ipynb <==
import pandas as pd
import numpy as np
import os

# Load the data

# Set to True to force a full refresh of the data
full_refresh = True

repo_root = os.popen('git rev-parse --show-toplevel').read().strip()

data_folder = os.path.join(repo_root, 'multi-parameter-estimation', 'data')

# Get list of data directories
data_dirs = os.listdir(data_folder)
data_dirs = [d for d in data_dirs if os.path.isdir(os.path.join(data_folder, d))]

# skip old-data
if 'old-data' in data_dirs:
    data_dirs.remove('old-data')

data_dirs.sort()
data_dirs

new_data_dirs = data_dirs.copy()

if not full_refresh:
    for d in data_dirs:
        if os.path.exists(os.path.join(data_folder, d, "scaled_coincidences.csv")):
            new_data_dirs.remove(d)

new_data_dirs

def load_coincidences(data_dir):
    coincidences = pd.read_csv(os.path.join(data_folder, data_dir, "labelled_coincidences.csv"))
    coincidences["data_dir"] = data_dir
    return coincidences

coincidences_df = pd.concat([load_coincidences(d) for d in new_data_dirs], ignore_index=True)
coincidences_df

# Interleave the different tomography settings

first_tags_in_tomo_settings = coincidences_df.groupby(["data_dir", "tomography_setting_t", "tomography_setting_r"])["detector_a_time_tag"].first().reset_index()
first_tags_in_tomo_settings

coincidences_df["first_tag_in_tomo_setting"] = coincidences_df.merge(
    first_tags_in_tomo_settings,
    on=["data_dir", "tomography_setting_t", "tomography_setting_r"],
    how="left",
    suffixes=("", "_first")
)["detector_a_time_tag_first"]
coincidences_df["relative_time_tag"] = coincidences_df["detector_a_time_tag"] - coincidences_df["first_tag_in_tomo_setting"]
coincidences_df = coincidences_df.drop(columns=["first_tag_in_tomo_setting"])
coincidences_df = coincidences_df.sort_values(by=["data_dir", "relative_time_tag"])
coincidences_df = coincidences_df.reset_index(drop=True)
coincidences_df

# Remove half the non-double bunched events

# count the number of non-double-bunched events
non_double_bunched_count = coincidences_df[coincidences_df["estimation_label"].isin(["C", "SB"])].count()["estimation_label"]
number_to_remove = int(non_double_bunched_count * 0.5)

mask = np.zeros(len(coincidences_df), dtype=bool)
mask[coincidences_df["estimation_label"].isin(["C", "SB"])] = True
np.random.seed(42)  # For reproducibility
indices_to_remove = np.random.choice(np.where(mask)[0], size=number_to_remove, replace=False)
coincidences_df = coincidences_df.drop(indices_to_remove).reset_index(drop=True)
coincidences_df

# Chunk into groups of 40

# group every 40 coincidences together
coincidences_df["group"] = (coincidences_df.index // 40) + 1
coincidences_df = coincidences_df.sort_values(by=["data_dir", "group", "relative_time_tag"])
coincidences_df = coincidences_df.reset_index(drop=True)
coincidences_df

# Sum by estimation_label within each group
grouped_coincidences = coincidences_df.groupby(
    ["data_dir", "group", "estimation_label"]
).agg(
    {
        "relative_time_tag": "min",
        "detector_a_time_tag": "count"
    }
).reset_index()

# pivot the estimation_label to columns
grouped_coincidences = grouped_coincidences.pivot_table(
    index=["data_dir", "group",],
    columns="estimation_label",
    values="detector_a_time_tag",
    fill_value=0
).reset_index()
grouped_coincidences.columns.name = None 
# ensure the columns all exist
for label in ["C", "SB", "DB_V"]:
    if label not in grouped_coincidences.columns:
        grouped_coincidences[label] = 0
grouped_coincidences["N"] = grouped_coincidences["C"] + grouped_coincidences["SB"] + grouped_coincidences["DB_V"] + grouped_coincidences["DB_H"]
grouped_coincidences = grouped_coincidences[grouped_coincidences["N"] == 40].drop(columns=["group"])
grouped_coincidences

from concurrent.futures import ThreadPoolExecutor

grouped = dict(tuple(grouped_coincidences.groupby("data_dir")))

def save_group(data_dir):
    df_subset = grouped[data_dir]
    output_file = os.path.join(data_folder, data_dir, "chunked_coincidences_n=40.csv")
    df_subset.to_csv(output_file, index=False)
    print(f"Saved {output_file}")

with ThreadPoolExecutor(max_workers=8) as executor:
    executor.map(save_group, new_data_dirs)


## Using the corrected labels

# Sum by estimation_label within each group
grouped_corrected_coincidences = coincidences_df.groupby(
    ["data_dir", "group", "corrected_estimation_label"]
).agg(
    {
        "relative_time_tag": "min",
        "detector_a_time_tag": "count"
    }
).reset_index()

# pivot the estimation_label to columns
grouped_corrected_coincidences = grouped_corrected_coincidences.pivot_table(
    index=["data_dir", "group",],
    columns="corrected_estimation_label",
    values="detector_a_time_tag",
    fill_value=0
).reset_index()
grouped_corrected_coincidences.columns.name = None 
# ensure the columns all exist
for label in ["C'", "SB", "HH", "VV"]:
    if label not in grouped_corrected_coincidences.columns:
        grouped_corrected_coincidences[label] = 0
grouped_corrected_coincidences["N"] = grouped_corrected_coincidences["C'"] + grouped_corrected_coincidences["SB"] + grouped_corrected_coincidences["HH"] + grouped_corrected_coincidences["VV"]
grouped_corrected_coincidences = grouped_corrected_coincidences[grouped_corrected_coincidences["N"] == 40].drop(columns=["group"])
grouped_corrected_coincidences

grouped = dict(tuple(grouped_corrected_coincidences.groupby("data_dir")))

def save_group(data_dir):
    df_subset = grouped[data_dir]
    output_file = os.path.join(data_folder, data_dir, "corrected_chunked_coincidences_n=40.csv")
    df_subset.to_csv(output_file, index=False)
    print(f"Saved {output_file}")

with ThreadPoolExecutor(max_workers=8) as executor:
    executor.map(save_group, new_data_dirs)


# Create higher n estimates

grouped_coincidences["group_80"] = (grouped_coincidences.index // 2) + 1
grouped_coincidences["group_120"] = (grouped_coincidences.index // 3) + 1
grouped_coincidences["group_160"] = (grouped_coincidences.index // 4) + 1
grouped_coincidences["group_200"] = (grouped_coincidences.index // 5) + 1

# sum by group_80, group_120, group_160, group_200
grouped_coincidences_80 = grouped_coincidences.groupby(
    ["data_dir", "group_80"]
).agg(
    {
        "C": "sum",
        "SB": "sum",
        "DB_V": "sum",
        "DB_H": "sum",
        "N": "sum"
    }
).reset_index()

grouped_coincidences_120 = grouped_coincidences.groupby(
    ["data_dir", "group_120"]
).agg(
    {
        "C": "sum",
        "SB": "sum",
        "DB_V": "sum",
        "DB_H": "sum",
        "N": "sum"
    }
).reset_index()

grouped_coincidences_160 = grouped_coincidences.groupby(
    ["data_dir", "group_160"]
).agg(
    {
        "C": "sum",
        "SB": "sum",
        "DB_V": "sum",
        "DB_H": "sum",
        "N": "sum"
    }
).reset_index()

grouped_coincidences_200 = grouped_coincidences.groupby(
    ["data_dir", "group_200"]
).agg(
    {
        "C": "sum",
        "SB": "sum",
        "DB_V": "sum",
        "DB_H": "sum",
        "N": "sum"
    }
).reset_index()

grouped_coincidences_80 = grouped_coincidences_80[grouped_coincidences_80["N"] == 80].drop(columns=["group_80"])
grouped_coincidences_120 = grouped_coincidences_120[grouped_coincidences_120["N"] == 120].drop(columns=["group_120"])
grouped_coincidences_160 = grouped_coincidences_160[grouped_coincidences_160["N"] == 160].drop(columns=["group_160"])
grouped_coincidences_200 = grouped_coincidences_200[grouped_coincidences_200["N"] == 200].drop(columns=["group_200"])

grouped_coincidences_200

def save_grouped_data(df, n):
    df = df[df["N"] == n]
    grouped = dict(tuple(df.groupby("data_dir")))

    def save_group(data_dir):
        df_subset = grouped[data_dir]
        output_file = os.path.join(data_folder, data_dir, f"chunked_coincidences_n={n}.csv")
        df_subset.to_csv(output_file, index=False)
        print(f"Saved {output_file}")

    with ThreadPoolExecutor(max_workers=8) as executor:
        executor.map(save_group, new_data_dirs)


save_grouped_data(grouped_coincidences_80, 80)
save_grouped_data(grouped_coincidences_120, 120)
save_grouped_data(grouped_coincidences_160, 160)
save_grouped_data(grouped_coincidences_200, 200)

## Using corrected labels

grouped_corrected_coincidences["group_80"] = (grouped_corrected_coincidences.index // 2) + 1
grouped_corrected_coincidences["group_120"] = (grouped_corrected_coincidences.index // 3) + 1
grouped_corrected_coincidences["group_160"] = (grouped_corrected_coincidences.index // 4) + 1
grouped_corrected_coincidences["group_200"] = (grouped_corrected_coincidences.index // 5) + 1

# sum by group_80, group_120, group_160, group_200
grouped_corrected_coincidences_80 = grouped_corrected_coincidences.groupby(
    ["data_dir", "group_80"]
).agg(
    {
        "C'": "sum",
        "SB": "sum",
        "VV": "sum",
        "HH": "sum",
        "N": "sum"
    }
).reset_index()

grouped_corrected_coincidences_120 = grouped_corrected_coincidences.groupby(
    ["data_dir", "group_120"]
).agg(
    {
        "C'": "sum",
        "SB": "sum",
        "VV": "sum",
        "HH": "sum",
        "N": "sum"
    }
).reset_index()

grouped_corrected_coincidences_160 = grouped_corrected_coincidences.groupby(
    ["data_dir", "group_160"]
).agg(
    {
        "C'": "sum",
        "SB": "sum",
        "VV": "sum",
        "HH": "sum",
        "N": "sum"
    }
).reset_index()

grouped_corrected_coincidences_200 = grouped_corrected_coincidences.groupby(
    ["data_dir", "group_200"]
).agg(
    {
        "C'": "sum",
        "SB": "sum",
        "VV": "sum",
        "HH": "sum",
        "N": "sum"
    }
).reset_index()

grouped_corrected_coincidences_80 = grouped_corrected_coincidences_80[grouped_corrected_coincidences_80["N"] == 80].drop(columns=["group_80"])
grouped_corrected_coincidences_120 = grouped_corrected_coincidences_120[grouped_corrected_coincidences_120["N"] == 120].drop(columns=["group_120"])
grouped_corrected_coincidences_160 = grouped_corrected_coincidences_160[grouped_corrected_coincidences_160["N"] == 160].drop(columns=["group_160"])
grouped_corrected_coincidences_200 = grouped_corrected_coincidences_200[grouped_corrected_coincidences_200["N"] == 200].drop(columns=["group_200"])

grouped_corrected_coincidences_200

def save_grouped_data(df, n):
    df = df[df["N"] == n]
    grouped = dict(tuple(df.groupby("data_dir")))

    def save_group(data_dir):
        df_subset = grouped[data_dir]
        output_file = os.path.join(data_folder, data_dir, f"corrected_chunked_coincidences_n={n}.csv")
        df_subset.to_csv(output_file, index=False)
        print(f"Saved {output_file}")

    with ThreadPoolExecutor(max_workers=8) as executor:
        executor.map(save_group, new_data_dirs)


save_grouped_data(grouped_corrected_coincidences_80, 80)
save_grouped_data(grouped_corrected_coincidences_120, 120)
save_grouped_data(grouped_corrected_coincidences_160, 160)
save_grouped_data(grouped_corrected_coincidences_200, 200)
