==> multi-parameter-estimation/post-processing/4.1_estimate_on_corrected_chunks.ipynb <==
import pandas as pd
import numpy as np
import os

# Set to True to force a full refresh of the data
full_refresh = True
eta = 0.968

# Load data

repo_root = os.popen('git rev-parse --show-toplevel').read().strip()

data_folder = os.path.join(repo_root, 'multi-parameter-estimation', 'data')

# Get list of data directories
data_dirs = os.listdir(data_folder)
data_dirs = [d for d in data_dirs if os.path.isdir(os.path.join(data_folder, d))]

# skip old-data
if 'old-data' in data_dirs:
    data_dirs.remove('old-data')

data_dirs.sort()
data_dirs

new_data_dirs = data_dirs.copy()

if not full_refresh:
    for d in data_dirs:
        if os.path.exists(os.path.join(data_folder, d, "estimators.csv")):
            new_data_dirs.remove(d)

new_data_dirs

def load_chunks(data_dir, n):
    if not os.path.exists(os.path.join(data_folder, data_dir, f"corrected_chunked_coincidences_n={n}.csv")):
        print(f"Skipping {data_dir} n={n} as file does not exist.")
        return pd.DataFrame()
    coincidences = pd.read_csv(os.path.join(data_folder, data_dir, f"corrected_chunked_coincidences_n={n}.csv"))
    coincidences["data_dir"] = data_dir
    return coincidences

chunks_df = pd.concat([pd.concat([load_chunks(d, n) for d in new_data_dirs]) for n in [40,80,120,160,200]], ignore_index=True)
chunks_df

# All samples

all_samples = chunks_df[chunks_df["N"] == 40].groupby("data_dir").sum().reset_index()

# append the new data to the existing data
chunks_df = pd.concat([chunks_df, all_samples], ignore_index=True)
chunks_df

# Estimate

chunks_df["theta_estimate"] = np.arccos(
    (chunks_df["HH"] - chunks_df["VV"]) / chunks_df["N"]
)

chunks_df

chunks_df["delta_phi_estimate"] = (
    1 / 2
    * np.arccos(
        1
        / eta
        * (chunks_df["SB"] - chunks_df["C'"])
        / (chunks_df["SB"] + chunks_df["C'"])
    )
)

chunks_df

chunks_df["clipped_delta_phi_estimate"] = (
    1 / 2
    * np.arccos(
        np.clip(
        1
        / eta
        * (chunks_df["SB"] - chunks_df["C'"])
        / (chunks_df["SB"] + chunks_df["C'"]), -1, 1)
    )
)

chunks_df

chunks_df[chunks_df["data_dir"] == "2025-06-03--20h-15m-22s"].count()

# Save

from concurrent.futures import ThreadPoolExecutor
import os

# Group chunks_df by data_dir for efficient access
grouped_chunks = dict(tuple(chunks_df.groupby("data_dir")))

def save_estimators(data_dir):
    if data_dir in grouped_chunks:
        df_subset = grouped_chunks[data_dir]
        output_file = os.path.join(data_folder, data_dir, "corrected_estimators.csv")
        df_subset.to_csv(output_file, index=False)
        print(f"Saved estimators for {data_dir} to {output_file}")
    else:
        print(f"No data to save for {data_dir}")

# Use threading to speed up saving
with ThreadPoolExecutor(max_workers=8) as executor:
    executor.map(save_estimators, new_data_dirs)
